ft: make changes according to the review
