+++ src/App.tsx
@@ -2,21 +2,9 @@ import { useEffect, useState } from "react";
 import { Element } from "./components/Element";
 import { WinCount } from "./components/WinCount";
 import { Result } from "./components/Result";
-import { elementTypes, evalResult, WinType } from "./Utils";
-
-
-function generateComputerAnswer(): elementTypes {
-  const randomNum = Math.random() * 1000; 
-
-  if(randomNum > 650){
-    return elementTypes.ROCK;
-  } else if(randomNum > 350){
-    return elementTypes.PAPER;
-  } else {
-    return elementTypes.SCISSOR;
-  }
-}
-
+import { elementTypes, WinType } from "./lib/types/types";
+import { evalResult, generateComputerAnswer } from "./lib/helpers/utils";
+import { elemnts } from "./lib/config/config";
 
 function App() {
   const [winner, setWinner] = useState<WinType | undefined>();
@@ -38,25 +26,28 @@ function App() {
 
   }, [compSelection]);
 
-  const handleUserSelect = (type: elementTypes) => {
+  function handleUserSelect(type: elementTypes){
       setUserSelection(type)
   };
 
-  const handleReset = () => {
+  function handleReset() {
     setUserSelection(undefined);
     setCompSelection(undefined);
     setWinner(undefined);
     setReset(true);
   }
+
     return (
       <>
         <h2 className="title">Scissor paper rock</h2>
-        <Element type="Rock" userSelects={handleUserSelect} computerChoice={compSelection} reset={reset}/>
+        {elemnts.map(element => {
+          return <Element type={element.type} userSelects={handleUserSelect} computerChoice={compSelection} reset={reset}/>
+        })}
+        {/* <Element type="Rock" userSelects={handleUserSelect} computerChoice={compSelection} reset={reset}/>
         <Element type="Paper" userSelects={handleUserSelect} computerChoice={compSelection} reset={reset}/>
-        <Element type="Scissor" userSelects={handleUserSelect} computerChoice={compSelection} reset={reset}/>
+        <Element type="Scissor" userSelects={handleUserSelect} computerChoice={compSelection} reset={reset}/> */}
         <WinCount winner={winner} />
-
-        <Result  winner={winner} userChoice={userSelection} computerChoice={compSelection}/>
+        {winner ? <Result  winner={winner} userChoice={userSelection} computerChoice={compSelection}/> : null}       
         <button onClick={handleReset}>Try Again</button>
       </>
     );
