+++ src/components/Result.tsx
@@ -6,16 +6,13 @@ type ResultProps = {
     winner: WinType | undefined,
 }
 function Result({userChoice, computerChoice, winner}: ResultProps){
-    if(winner) {
-        return (
-            <div>
-                <p>You choose: {userChoice}</p>
-                <p>Computer choose: {computerChoice}</p>
-                <p>{winner == WinType.DRAW ? 'Draw' : `${winner} won`} </p>
-            </div>
-        )
-    } else return null;
-
+    return (
+        <div>
+            <p>You choose: {userChoice}</p>
+            <p>Computer choose: {computerChoice}</p>
+            <p>{winner == WinType.DRAW ? 'Draw' : `${winner} won`} </p>
+        </div>
+    )
 }
 
 export { Result };
