+++ src/lib/config/config.ts
@@ -1,0 +1,7 @@
+import { elementTypes } from "../types/types";
+
+export const elemnts = [
+    {type: elementTypes.ROCK},
+    {type: elementTypes.PAPER},
+    {type: elementTypes.SCISSOR},
+]
