+++ src/lib/types/types.ts
@@ -1,0 +1,11 @@
+export enum elementTypes {
+    ROCK = 'Rock',
+    PAPER = 'Paper',
+    SCISSOR = 'Scissor',
+  }
+  
+  export enum WinType {
+    USER = 'User',
+    COMPUTER = 'Computer',
+    DRAW = 'Draw'
+  }
